generic list + text export

+++ generator/flowExtractor/flows.h
@@ -93,7 +93,7 @@ void clear_hash_recordv6(flowv6_record**  hash_table);
 void display_flowv4(flowv4_record* flow, FILE* fptr, bool normal);
 void display_flowv6(flowv6_record* flow);
 void update_stats(flowv4_record* record, uint16_t size, struct timeval ts);
-uint32_t compute_inter_arrival(struct timeval* t1, struct timeval* t2);
+float compute_inter_arrival(struct timeval* t1, struct timeval* t2);
 void export_flowv4_to_file(flowv4_record* flow, FILE* fptr);
 void export_binary_flowv4_to_file(flowv4_record* flow, FILE* fptr);
 void export_allv4_to_file(flowv4_record** hash_table, FILE* fptr);
